refactor: updated interface file names
