+++ src/core/application/useCases/produto/produto.interface.ts
