+++ src/core/domain/repositories/produtoRepository.interface.ts
